took the combobox button out of tab order
https://www.w3.org/TR/wai-aria-practices-1.1/#combobox
For each combobox pattern the button need not be in the tab order if
there is an appropriate keystroke associated with the input element
such that when focus is on the input, the keystroke triggers display of
the associated drop down list. In this case, a Tab to focus the button
is unnecessary. This is the same behavior as the select element.